requirements.txt
streamlit
pillow
pytesseract
googletrans==4.0.0-rc1
numpy